FileExplorer3: Added Explorer Zoom (via ExplorerVM.UIScale) and FileList Zoom (via ViewModeCommand.SliderValue in FileListVM.Commands.ToolbarCommands), only usable using Control+Mousewheel (pinch is not supportted). FileExplorer3: FileList/ExplorerCommandManager.ScriptCommands.ZoomIn/Out property now point to Explorer/FileList.Zoom() IScriptCommand, and NavigationCommands.In/DecreaseZoom in FileList/Explorer CommandManager now point to ScriptCommands.ZoomIn/Out.

+++ src/FileExplorer3/app/FileExplorer3/UserControls/ListViewEx.cs
@@ -215,7 +215,7 @@ namespace FileExplorer.UserControls
         /// </summary>
         public double ItemSize
         {
-            get { return (int)GetValue(ItemSizeProperty); }
+            get { return (double)GetValue(ItemSizeProperty); }
             set { SetValue(ItemSizeProperty, value); }
         }
